Removing the packet size limitation

+++ trace_driven_simulator/internal/simulator/models.go
@@ -21,7 +21,7 @@ const (
 )
 
 type Packet struct {
-	Size             uint16
+	Size             uint32
 	ArrivalTime      float32
 	StartServiceTime float32
 	DepartureTime    float32
